Make sure the widnode time is given in unicode
The value returned by time.strftime is encoded depending on the locale. This
causes the creation on the loca variable newEntry into the function addEntry
function to fail if the widnode.time contains no ascii encodable characters.

Transforming it in unicode is the recommended way of doing in python and avoids
such problems.

See
http://stackoverflow.com/questions/19412915/how-determine-encoding-of-datetime-strftime-in-python
for more details